Moved GraphType into its own file and toyed around with some helper functions.
