+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/Registry/RegistryPlaceholder.cs
@@ -177,192 +177,5 @@ namespace UnityEditor.ShaderGraph.Registry
                 return shaderFunctionBuilder.Build();
             }
         }
-
-        internal class GraphType : Defs.ITypeDefinitionBuilder
-        {
-            public static RegistryKey kRegistryKey => new RegistryKey { Name = "GraphType", Version = 1 };
-            public RegistryKey GetRegistryKey() => kRegistryKey;
-            public RegistryFlags GetRegistryFlags() => RegistryFlags.Type;
-
-            public enum Precision { Fixed, Half, Single, Any }
-            public enum Primitive { Bool, Int, Float, Any }
-            public enum Length { One = 1, Two = 2, Three = 3, Four = 4, Any = -1 }
-            public enum Height { One = 1, Two = 2, Three = 3, Four = 4, Any = -1 }
-            public enum Usage { In, Out, Static, Local }
-
-            // Values here represent a resolving priority.
-            // The highest numeric value has the highest priority.
-            public static readonly Dictionary<Precision, int> PrecisionToPriority = new()
-            {
-                { Precision.Fixed, 1 },
-                { Precision.Half, 2 },
-                { Precision.Single, 3 },
-                { Precision.Any, -1}
-            };
-            public static readonly Dictionary<Primitive, int> PrimitiveToPriority = new()
-            {
-                { Primitive.Bool, 1 },
-                { Primitive.Int, 2 },
-                { Primitive.Float, 3 },
-                { Primitive.Any, -1 }
-            };
-            public static readonly Dictionary<Length, int> LengthToPriority = new()
-            {
-                { Length.One, 1 },
-                { Length.Two, 4 },
-                { Length.Three, 3 },
-                { Length.Four, 2 },
-                { Length.Any, -1 }
-            };
-            public static readonly Dictionary<Height, int> HeightToPriority = new()
-            {
-                { Height.One, 1 },
-                { Height.Two, 4 },
-                { Height.Three, 3 },
-                { Height.Four, 2 },
-                { Height.Any, -1 }
-            };
-
-
-            public const string kPrimitive = "Primitive";
-            public const string kPrecision = "Precision";
-            public const string kLength = "Length";
-            public const string kHeight = "Height";
-            public const string kEntry = "_Entry";
-
-            public void BuildType(IFieldReader userData, IFieldWriter typeWriter, Registry registry)
-            {
-                // default initialize to a float4.
-                typeWriter.SetField(kPrecision, Precision.Single);
-                typeWriter.SetField(kPrimitive, Primitive.Float);
-                typeWriter.SetField(kLength, Length.Four);
-                typeWriter.SetField(kHeight, Height.One);
-
-                // read userdata and make sure we have enough fields.
-                if (!userData.GetField(kLength, out Length length))
-                    length = Length.Four;
-                if (!userData.GetField(kHeight, out Height height))
-                    height = Height.One;
-
-                // ensure that enough subfield values exist to represent userdata's current data.
-                for (int i = 0; i < (int)length * (int)height; ++i)
-                    typeWriter.SetField<float>($"c{i}", 0);
-            }
-
-            string Defs.ITypeDefinitionBuilder.GetInitializerList(IFieldReader data, Registry registry)
-            {
-                data.GetField(kLength, out Length length);
-                data.GetField(kHeight, out Height height);
-                int l = Mathf.Clamp((int)length, 1, 4);
-                int h = Mathf.Clamp((int)height, 1, 4);
-
-                string result = $"{((Defs.ITypeDefinitionBuilder)this).GetShaderType(data, new ShaderFoundry.ShaderContainer(), registry).Name}" + "(";
-                for(int i = 0; i < l * h; ++i)
-                {
-                    data.GetField($"c{i}", out float componentValue);
-                    result += $"{componentValue}";
-                    if (i != l * h - 1)
-                        result += ", ";
-                }
-                result += ")";
-                return result;
-            }
-
-            ShaderFoundry.ShaderType Defs.ITypeDefinitionBuilder.GetShaderType(IFieldReader data, ShaderFoundry.ShaderContainer container, Registry registry)
-            {
-                data.GetField(kPrimitive, out Primitive primitive);
-                data.GetField(kPrecision, out Precision precision);
-                data.GetField(kLength, out Length length);
-                data.GetField(kHeight, out Height height);
-                int l = Mathf.Clamp((int)length, 1, 4);
-                int h = Mathf.Clamp((int)height, 1, 4);
-
-                string name = "float";
-
-                switch(primitive)
-                {
-                    case Primitive.Bool: name = "bool"; break;
-                    case Primitive.Int: name = "int"; break;
-                    case Primitive.Float:
-                        switch (precision)
-                        {
-                            case Precision.Fixed: name = "fixed"; break;
-                            case Precision.Half: name = "half"; break;
-                        }
-                        break;
-                }
-
-                var shaderType = ShaderFoundry.ShaderType.Scalar(container, name);
-
-                if (h != 1 && l != 1)
-                {
-                    shaderType = ShaderFoundry.ShaderType.Matrix(container, shaderType, l, h);
-                }
-                else
-                {
-                    shaderType = ShaderFoundry.ShaderType.Vector(container, shaderType, Mathf.Max(l, h));
-                }
-                return shaderType;
-            }
-        }
-
-        internal class GraphTypeAssignment : Defs.ICastDefinitionBuilder
-        {
-            public RegistryKey GetRegistryKey() => new RegistryKey { Name = "GraphTypeAssignment", Version = 1 };
-            public RegistryFlags GetRegistryFlags() => RegistryFlags.Cast;
-            public (RegistryKey, RegistryKey) GetTypeConversionMapping() => (kRegistryKey, kRegistryKey);
-            public bool CanConvert(IFieldReader src, IFieldReader dst) => true;
-
-
-            private static string MatrixCompNameFromIndex(int i, int d)
-            {
-                return $"_mm{ i / d }{ i % d }";
-            }
-            private static string VectorCompNameFromIndex(int i)
-            {
-                switch(i)
-                {
-                    case 0: return "x";
-                    case 1: return "y";
-                    case 2: return "z";
-                    default: return "w";
-                }
-            }
-
-
-            ShaderFoundry.ShaderFunction Defs.ICastDefinitionBuilder.GetShaderCast(IFieldReader src, IFieldReader dst, ShaderFoundry.ShaderContainer container, Registry registry)
-            {
-                // In this case, we can determine a casting operation purely from the built types. We don't actually need to analyze field data.
-                // We will get precision truncation warnings though...
-                var srcType = registry.GetTypeBuilder(src.GetRegistryKey()).GetShaderType(src, container, registry);
-                var dstType = registry.GetTypeBuilder(dst.GetRegistryKey()).GetShaderType(dst, container, registry);
-
-                string castName = $"Cast{srcType.Name}_{dstType.Name}";
-                var builder = new ShaderFoundry.ShaderFunction.Builder(container, castName);
-                builder.AddInput(srcType, "In");
-                builder.AddOutput(dstType, "Out");
-
-                var srcSize = srcType.IsVector ? srcType.VectorDimension : srcType.IsMatrix ? srcType.MatrixColumns * srcType.MatrixRows : 1;
-                var dstSize = dstType.IsVector ? dstType.VectorDimension : dstType.IsMatrix ? dstType.MatrixColumns * dstType.MatrixRows : 1;
-
-                string body = $"Out = {srcType.Name} {{ ";
-
-                for (int i = 0; i < dstSize; ++i)
-                {
-                    if (i < srcSize)
-                    {
-                        if (dstType.IsMatrix) body += $"In.{MatrixCompNameFromIndex(i, dstType.MatrixColumns)}"; // are we row or column major?
-                        if (dstType.IsVector) body += $"In.{VectorCompNameFromIndex(i)}";
-                        if (dstType.IsScalar) body += $"In";
-                    }
-                    else body += "0";
-                    if (i != dstSize - 1) body += ", ";
-                }
-                body += " };";
-
-                builder.AddLine(body);
-                return builder.Build();
-            }
-        }
     }
 }
