+++ com.unity.shadergraph/Editor/GraphDeltaRegistry/StandardDefinitions/GraphType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b5f184052962b94da5e3940ebcbead9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
